+++ components/merkledb/src/views/address.rs
@@ -150,7 +150,7 @@ impl IndexAddress {
     }
 
     /// Full address with a separator between `name` and `bytes` represented as byte array.
-    pub(super) fn fully_qualified_name(&self) -> Vec<u8> {
+    pub(crate) fn fully_qualified_name(&self) -> Vec<u8> {
         /// Separator between the name and the additional bytes in family indexes.
         const INDEX_NAME_SEPARATOR: &[u8] = &[SEPARATOR_CHAR];
         const MIGRATION_PREFIX: &[u8] = &[MIGRATION_CHAR];
@@ -163,6 +163,15 @@ impl IndexAddress {
             (true, None) => concat_keys!(MIGRATION_PREFIX, self.name()),
             (false, None) => self.name.as_bytes().to_vec(),
         }
+    }
+
+    /// Returns the common prefix of fully qualified names for the child indexes.
+    pub(crate) fn qualified_prefix(&self) -> Vec<u8> {
+        let mut prefix = self.fully_qualified_name();
+        if self.id_in_group.is_none() {
+            prefix.push(SEPARATOR_CHAR);
+        }
+        prefix
     }
 
     /// Infers the name part of the fully qualified name that was obtained with